Add Clash.Normalize.Transformations.Case
All transformations which are exclusively concerning case expressions
now live in a separate module. This is part of work to make the
normalization modules quicker to compile.
