+++ clash-lib/clash-lib.cabal
@@ -242,6 +242,7 @@ Library
                       Clash.Normalize.PrimitiveReductions
                       Clash.Normalize.Strategy
                       Clash.Normalize.Transformations
+                      Clash.Normalize.Transformations.Case
                       Clash.Normalize.Types
                       Clash.Normalize.Util
